PMM-13795: External Services tests
